ok so its a sloppy first version of import csv
need to go through import logic again some time

if im going to use javers. well, use it all the way.
make aspects on javers instead of repositories.

still missing data. not sure where to find it and so on
now atleast i got some cows in the db! *yihaa*