burn... kafka was running in the background
so didnt realize the test now are dependent on it.
need to think about how to handle testing.
i can see that jhipster already has added the dependency
org.testcontainers.kafka
but i will probably need to set it up myself... or??

in order to go forth EntityChangeJaversAspect is only
active in dev and prod